Added minimal support for spin polarized calculations
This converts a spin-polarized calculation from QE into a spinor structure in INTW through pw2intw, to be used as a standard spinor calculation in the subsequent INTW calculations. Not the most efficient but easiest to maintain.

Still experimental, needs more testing, but does not affect any of the standard scalar or spinor calculations, well tested before.

Also corrected an allocation bug when TR_symmetry=.false.